+++ apps/src/lab2/views/components/Instructions/ResourcePanel/productTours/onboardingTourSteps.ts
