+++ apps/src/lab2/views/components/Instructions/ResourcePanel/productTours/validationTourSteps.ts
