+++ apps/src/sketchlab/useSketchlabTour.ts
@@ -8,7 +8,7 @@ import useStartTourWhenAvailable from '@cdo/apps/sharedComponents/productTour/us
 import {tryGetLocalStorage} from '@cdo/apps/utils';
 
 import {SKETCHLAB_ONBOARDING_TOUR_SEEN} from './constants';
-import {createSketchlabTourSteps} from './sketchlabShepherdTourSteps';
+import {createSketchlabTourSteps} from './sketchlabTourSteps';
 
 const SKETCHLAB_ONBOARDING_FLOW_NAME = 'Sketch Lab Onboarding V2';
 
@@ -28,7 +28,7 @@ const onTourCancel = (stepIndex: number) =>
     step: stepIndex.toString(),
   });
 
-const useSketchlabShepherdTour = () => {
+const useSketchlabTour = () => {
   // Wait for the Excalidraw toolbar to be fully rendered before starting the tour.
   const [isToolbarReady, setIsToolbarReady] = useState(false);
   useEffect(() => {
@@ -83,4 +83,4 @@ const useSketchlabShepherdTour = () => {
   useStartTourWhenAvailable(tour);
 };
 
-export default useSketchlabShepherdTour;
+export default useSketchlabTour;
